de-dup ipv4 and ipv6 function implementation
