+++ rust/connlib/tunnel/src/device_channel.rs
@@ -22,7 +22,7 @@ use crate::ip_packet::{IpPacket, MutableIpPacket};
 use connlib_shared::error::ConnlibError;
 use connlib_shared::messages::Interface;
 use connlib_shared::{Callbacks, Error};
-use ip_network::IpNetwork;
+use ip_network::{IpNetwork, Ipv4Network, Ipv6Network};
 use pnet_packet::Packet;
 use std::collections::HashSet;
 use std::io;
@@ -36,6 +36,22 @@ pub struct Device {
     tun: Option<Tun>,
     waker: Option<Waker>,
     mtu_refreshed_at: Instant,
+}
+
+#[allow(dead_code)]
+fn ipv4(ip: &IpNetwork) -> Option<&Ipv4Network> {
+    match ip {
+        IpNetwork::V4(v4) => Some(v4),
+        IpNetwork::V6(_) => None,
+    }
+}
+
+#[allow(dead_code)]
+fn ipv6(ip: &IpNetwork) -> Option<&Ipv6Network> {
+    match ip {
+        IpNetwork::V4(_) => None,
+        IpNetwork::V6(v6) => Some(v6),
+    }
 }
 
 impl Device {
