+++ rust/connlib/tunnel/src/device_channel/tun_android.rs
@@ -1,4 +1,4 @@
-use crate::device_channel::ioctl;
+use crate::device_channel::{ioctl, ipv4, ipv6};
 use connlib_shared::{messages::Interface as InterfaceConfig, Callbacks, Error, Result};
 use ip_network::{IpNetwork, Ipv4Network, Ipv6Network};
 use std::net::IpAddr;
