+++ rust/connlib/tunnel/src/device_channel/tun_windows.rs
@@ -1,3 +1,4 @@
+use crate::device_channel::{ipv4, ipv6};
 use connlib_shared::{messages::Interface as InterfaceConfig, Result};
 use ip_network::IpNetwork;
 use std::{
